Made everything guint

+++ io.c
@@ -47,17 +47,18 @@ void setSocketTimeout(gint desc, gint milliseconds) {
   }
 }
 
-guint getMaxSeq(FILE* inputFile){
+guint getMaxSeq(FILE* inputFile) {
+  //TODO: Remove alerts
 	Datagram dgram = {0};
 	fseek (inputFile, 0, SEEK_END);
 	guint size = ftell(inputFile);
 	alert("size %u", size);
-	guint maxSeq = (size/sizeof(dgram.segment.data)) +1;
+	guint maxSeq = (size / sizeof(dgram.segment.data)) + 1;
 	alert("maxSeq: %u", maxSeq);
 	return maxSeq;
 }
 
-Datagram readInputData(FILE *inputFile, gsize seqNumber) {
+Datagram readInputData(FILE *inputFile, guint seqNumber) {
   //TODO: This might slow down IO, we could try using a bigger buffer or so
   Datagram dgram = {0};
   fseek(inputFile, (seqNumber - 1) * sizeof(dgram.segment.data), SEEK_SET);
